Add new card to practice inline/inline-block/span
The distinction between inline and inline-block display types is in
how margins and width are treated. In inline elements, the width is
constant - only as much as it needs, and vertical margins are not
respected. Inline-block elements, however, these properties apply.
These characteristics were tested with the new class selectors added.

A brief test for the span tag was used as practice.